Refactor non-glTF specific content into shared scatter file.
